Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,11 +2,11 @@
 
 
 int main() {
-    int a,b,c;
-    scanf("%d",&a);
-    scanf("%d",&b);
+    float a,b,c;
+    scanf("%f",&a);
+    scanf("%f",&b);
 
-    scanf("%d",&c);
+    scanf("%f",&c);
     float avg;
     avg=(a+b+c)/3;
     printf("Average: %.2f", avg);
